fix: Change averageClickPosition of the ABTestResponse.Variant type to Double

+++ Sources/AlgoliaSearchClient/Models/Analytics/ABTest/Response/ABTestResponse+Variant.swift
@@ -29,7 +29,7 @@ extension ABTestResponse {
     public let noResultCount: Int?
 
     /// Average click position for the variant.
-    public let averageClickPosition: Int?
+    public let averageClickPosition: Double?
 
     public let searchCount: Int?
